option for youtube collection interval
