+++ commstats/commstat.go
@@ -6,22 +6,16 @@ package commstats
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"github.com/raedahgroup/dcrextdata/app/config"
 	"net/http"
-	"strconv"
 	"time"
 
 	"github.com/raedahgroup/dcrextdata/app"
+	"github.com/raedahgroup/dcrextdata/app/config"
 )
 
 const (
 	dateTemplate     = "2006-01-02 15:04"
 	dateMiliTemplate = "2006-01-02 15:04:05.99"
-	youtubeChannelId = "UCJ2bYDaPYHpSmJPh_M5dNSg"
-	youtubeKey       = "AIzaSyBmUyMNtZUqReP2NTs39UlTjd9aUjXWKq0"
 	retryLimit       = 3
 )
 
@@ -146,80 +140,10 @@ func (c *Collector) collectAndStore(ctx context.Context) error {
 
 	go c.startTwitterCollector(ctx)
 
-	// youtube
-	youtubeSubscribers, err := c.getYoutubeSubscriberCount(ctx)
-	for retry := 0; err != nil; retry++ {
-		if retry == retryLimit {
-			return err
-		}
-		log.Warn(err)
-		youtubeSubscribers, err = c.getYoutubeSubscriberCount(ctx)
-	}
-
-	youtubeStat := Youtube{
-		Date:        time.Now().UTC(),
-		Subscribers: youtubeSubscribers,
-	}
-	err = c.dataStore.StoreYoutubeStat(ctx, youtubeStat)
-	if err != nil {
-		log.Error("Unable to save Youtube stat, %s", err.Error())
-		return err
-	}
-
-	log.Infof("New Youtube stat collected at %s, Subscribers %d",
-		youtubeStat.Date.Format(dateMiliTemplate), youtubeSubscribers)
+	go c.startYoutubeCollector(ctx)
 
 	// github
 	go c.startGithubCollector(ctx)
 
 	return nil
-}
-
-func (c *Collector) getYoutubeSubscriberCount(ctx context.Context) (int, error) {
-	if ctx.Err() != nil {
-		return 0, ctx.Err()
-	}
-
-	youtubeUrl := fmt.Sprintf("https://content.googleapis.com/youtube/v3/channels?key=%s&part=statistics&id=%s",
-		youtubeKey, youtubeChannelId)
-
-	request, err := http.NewRequest(http.MethodGet, youtubeUrl, nil)
-	if err != nil {
-		return 0, err
-	}
-
-	request.Header.Set("user-agent",
-		"Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/75.0.3770.100 Safari/537.36")
-
-	resp, err := c.client.Do(request.WithContext(ctx))
-	if err != nil {
-		return 0, err
-	}
-	defer resp.Body.Close()
-	var response struct {
-		Items []struct {
-			Statistics struct {
-				SubscriberCount string `json:"subscriberCount"`
-			} `json:"statistics"`
-		} `json:"items"`
-	}
-	if resp.StatusCode == http.StatusOK {
-		err = json.NewDecoder(resp.Body).Decode(&response)
-		if err != nil {
-			return 0, fmt.Errorf(fmt.Sprintf("Failed to decode json: %v", err))
-		}
-	} else {
-		return 0, fmt.Errorf("unable to fetch youtube subscribers: %s", resp.Status)
-	}
-
-	if len(response.Items) < 1 {
-		return 0, errors.New("unable to fetch youtube subscribers, no response")
-	}
-
-	subscribers, err := strconv.Atoi(response.Items[0].Statistics.SubscriberCount)
-	if err != nil {
-		return 0, errors.New("unable to fetch youtube subscribers, no response")
-	}
-
-	return subscribers, nil
 }
